Move Caer GUI into an appropriate folder
